requirements.txt
fastapi
uvicorn
playwright
Pillow
python-dotenv
supabase
